Change `/search/Index` to `/search/query/Index`
The routes for querying and index administration are ambiguous if
there is an index named `index`.  The user will not be able to query
because Yokozuna will interpret `GET /search/index?q=...` as a request
to the index admin resource.  This commit changes the query resource
to `/search/query/Index` so there is no ambiguity.