Fix failure with yz_dt_test due to a disable operation in a map
It is currently invalid to disable a flag in a newly created map with
no context. The use of disable on a flag is only valid after the map
has been created, written to Riak, and read from Riak. The context is
then established and the operation is able to complete. The fix for
the test for right now is to remove the disable operation on the flag
for the "Russell" map in the test so that no invalid operation are
attempted.