Prevent any Indexqs with 0-length queues (or batches in flight) from immediately going back into a batching state, when a drain operation is initiated.
This change also makes the yz_hashtree update occur synchronously within the context of the drain FSM, between the point in time that all data is drained from all solrqs, and the point in time when the solrqs are notified to resume batching.

Both of these are attempts to fix "leaky" indexqs, as well as to minimize the time window during which the yz and kv hash trees can diverge.