Use Base64 encoded names for fuses
Previously we had been directly using the name of each Solr index as the
name of the fuse that was associated with it. However, since the index
names are arbitrary names chosen by users, it is possible that these
names could clash or contain characters that are not legal for fuse
names. Therefore, we now apply a prefix to the fuse name to make sure
that it does not clash, and then apply Base64 encoding to that name to
ensure that it does not contain illegal characters.