Removed the explicit ibrowse settings we make in the code.
Some of these settings were made in 2.0.7 to work around some issues we discovered with the pipelining code in ibrowse-4.0.1, which may no longer be relevant.  We still need to do some performance and resource testing to be sure, before merging this change in.

There is another change in this commit to remove the inactivity_timeout setting that @rzezeski had added.  With the line not removed with this version of ibrowse, search requests time out after 60 seconds (default).

See https://github.com/basho/yokozuna/commit/b02c0bd with @rzezeski's rationale for adding this line.  We need to understand whether TIME_WAITs are still an issue in this version of ibrowse without this setting.