Throttle by entries sent to solrq_helper, rather than total queue size.
In order to set the throttle in _anticipation_ of load, rather than
reaction to it, we now set the throttle based on the average
number of entries _sent_ to a helper in a batch, rather than the number
of total entries in the queue. This is a much better predictor of
throughput drops (large batches slow us down) and seems to smooth out
performance fairly consistently.