Separated out asynchronous call to compare hashtrees and wrapped in a try-catch in order to prevent spawn-linked compare proc from crashing (and not replying back to the exchange FSM) when the riak_kv_index_hashtree being compared with is terminated.
Also addressed a bug which prevented exchanges from proceeding if
one exchange is wedged, by not immediately requeing the excahnge,
but instead waiting for the next full cycle of additions to the
exchange queue (on the yz_entropy_mgr).

removed compare_pid field, which wasn't really being used for anything.