Removed the verify_no_repair_for_non_indexed_data test.
The reasoning for the removal of this test is as follows:

Due to non-atomicity of snapshotting of the riak_kv AAE hash tree and the draining of solrqs, there is always a window of time where data can be written into a queue AFTER a KV hash tree has been snapshotted but before a drain initiates, resulting in data showing up in a YZ AAE hash tree but NOT in the KV snapshot.  This results in a "false" repair -- data showing up in Solr that is not in KV, which subsequently gets removed, and then re-added at the next repair for that partition.

We are tolerant of this race condition, as it is rare and will result in a small amount of spurious repair.  The test, therefore, is testing a condition we cannot guarantee, and the test does fail intermittently for this reason.  We have therefore decided to remove this test from the suite.