Performance fixes for Solr integration
- Remove Delete By Query calls whenever possible as it's a performance killer on Solr
- Split hashtree update out of the "Drain Timeout" window as we were killing hashtree builds unnecessarily.
- Fix issue where `yz_index_hashtree:fold_keys` was flooding the yz_index_hashtree mailbox by implementing the "90 async/1 sync" call pattern as other `index` and `delete` callers used.

Prior to this fix, whenever we would add a document to Solr that might contain siblings, we would add a delete by query operation to the batch we send to Solr (in a single HTTP request).  This had a significant impact on performance, especially when objects being updated had a lot of siblings.  (We found that deleting by query would contend for locks with the segment scheduler, and in general adding deletesByQuery to high workloads seems like a bad idea).

The fix involves changing deleteByQuery to deleting documents in Solr by their document ids, which has much better performance than deleting by query..  The challenge is that when adding an object, we didn't have the sibling vtags to generate a document id.  We therefore made a change to riak_kv (basho/riak_kv#1520) which, when updating an object, also sends the old object being replaced.  From that old object, we can extract any siblings and generate associated document ids to delete in Solr.

The process of doing a KV->Yokozuna hastree comparison takes several steps, including draining the relevant Solr queues before snapshotting the Yokozuna hashtree so that its state can be as close as possible to the KV hashtree. In order to not block the exchange indefinitely, a "Drain Timeout" was included in this process. However, as initially developed, the drain was not the only thing being measured in this timeout window. It also included the hashtree:update call, which, on a heavily-loaded system, can take several minutes. This would cause the "drain" to time out and be forcibly cancelled, even though the drain itself had only taken a few hundreds of milliseconds.

`yz_drain_fsm`, `yz_drain_mgr`,  and `yz_exchange_fsm` were was reworked to split the Solr drain from the hashtree update, resolving the issue of drain timeouts under normal operating conditions.

During performance testing, it was discovered that the `yz_index_hashtree:fold_keys`
function could swamp the mailbox of the `yz_index_hashtree` so that other processes
could not make progress. This commit makes `fold_keys` use the same
"90 async + 1 sync" call pattern that `yz_kv:index` used. That logic
(which lived in `yz_kv`) has been moved to `yz_index_hashtree` and
shared by the new `index` and `delete` calls that do not take an explicit
"call mode" parameter.
Future work is scheduled to figure out how to make this pattern more generic
(as it's used in several places) and to eliminate the use of the caller's
process dictionary for storing hashtree tokens.

* Added logging for clear and exchange trees, for audit of administrative operations.
* Modified the "Will Repair" log in the yz_exchange_fsm to track the "direction" of repair, viz., whether the repair resulted in a delete of Solr data or an add/update to Solr.
* Corrected the list of `required_queues` in `yz_solrq_sup` to include both _current_ and _future_ indexes the node owned. This prevented premature shutdown of a solrq worker/helper pair that was started by writes to a _future_ vnode (one in the process of being handed off to the node in question).
* Remove exometer statistics when an index is removed. Before, if an index was later re-added, the fuse creation would fail, eventually causing the node to crash. Test updated in 315b7464fa9996e56237de21050fd5f7b1f4eb6d
* Monitor `yz_solrq_drain_fsm` from the queues being drained in case the drain_fsm crashes. Before, it was possible for a queue to get stuck in `wait_for_drain_complete` state if the drain fsm crashed before the drain complete messages were sent.