Prevent fallback vnodes' solrqs from stopping if they are still needed
Before, we used the "current and next" vnode list according to the ring
to determine which partitions were needed for solrq start/stop. This
change simply asks `riak_core_vnode_manager` which `riak_kv_vnode`
vnodes are currently running, and uses that list to figure out which
queues we need to have running.
This should prevent issues where fallback vnodes currently running on a
node would create many short-lived solrqs, which would start on demand
when an index call was made that needed them, and stop every tick.