Use correct equality operator for sh
The shell scripts in the tools directory are executed via #!/usr/bin/env sh. Three of them--build-jar.sh, copy-jars.sh, and grab_solr.sh--use a bash idiom which is a double equals inside of a single bracket [ "$something" == "otherthing" ].

In bash == and = are equivalent inside of []. In sh that is not the case. Checking for equality inside of single brackets can only be done with a single equals sign. Using a double-equals results in an unexpected operator error, and the left hand side of the conditional statement always fails.